refactor server code add router/controller/service layers
